+++ src/main/kotlin/org/omegat/gradle/config/DefaultModule.kt
@@ -2,7 +2,7 @@ package org.omegat.gradle.config
 
 import org.gradle.api.Project
 import org.omegat.gradle.OmegatPlugin
-import java.util.*
+import java.util.Properties
 
 class DefaultModule(val project: Project) {
     private val props = Properties()
@@ -11,8 +11,8 @@ class DefaultModule(val project: Project) {
         props.load(OmegatPlugin::class.java.getResourceAsStream("omegat.properties"))
         project.repositories.jcenter()
         project.repositories.apply {
-            maven {
-                it.setUrl(props.getProperty("mavenRepositoryUrl"))
+            maven { artifactRepository ->
+                artifactRepository.setUrl(props.getProperty("mavenRepositoryUrl"))
             }
         }
     }