+++ src/main/kotlin/org/omegat/gradle/config/PluginExtension.kt
@@ -17,7 +17,7 @@ open class PluginExtension(val project: Project) {
     var pluginClass: String? = null
     var debugPort: Int? = null
 
-    val manifest: OmegatManifest = OmegatManifest(project, this)
+    val manifest: OmegatManifest = OmegatManifest(project)
 
     /**
      * Set the [packIntoJarFileFilter] with a Groovy [Closure]